collectors/velib/requirements.txt
Flask==3.*
gunicorn==22.*
requests==2.*
google-cloud-pubsub==2.*
google-auth==2.*